refactor : User DTO 패키지 분리
- User DTO 클래스를 dto 하위의 user 패키지로 이동합니다.
