+++ src/main/java/kr/codesqaud/cafe/dto/user/UserResponse.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.dto;
+package kr.codesqaud.cafe.dto.user;
 
 import kr.codesqaud.cafe.domain.User;
